+++ src/__tests__/lib/containers/Login.test.js
@@ -2,7 +2,7 @@ import React from 'react';
 import {configure, mount} from 'enzyme'
 import Adapter from 'enzyme-adapter-react-16';
 
-import Login from '../../../lib/containers/Login';
+import Login from '../../../lib/containers/demo_components/Login';
 
 configure({ adapter: new Adapter() });
