+++ src/lib/containers/demo_components/UserFavorites.js
